requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.7.1
python-dotenv==1.0.1
alembic==1.13.2
pytest==8.2.2
requests==2.32.3
